fix pr comments, fix test, move general wizard componets, fix design step1
