+++ src/components/Wizard/SectionBox.test.tsx
